+++ CHANGES.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added `IScope.ContainedScopes` which lets you check scopes contained within another scope.
 
 ## v0.2.7-beta.8
 ### Added